Moved from terminal to stdio output in arch.c, moved mb print code intto print.c file and added memory_info struct; Info from mb def!
